requirements.txt
annotated-types==0.7.0
blinker==1.8.2
blis==1.0.1
cachelib==0.9.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
cymem==2.0.8
dnspython==2.6.1
filelock==3.16.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Session==0.8.0
fsspec==2024.9.0
fuzzywuzzy==0.18.0
huggingface-hub==0.25.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
langcodes==3.4.1
language_data==1.2.0
Levenshtein==0.26.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.18.6
murmurhash==1.0.10
networkx==3.3
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
packaging==24.1
preshed==3.0.9
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.9.1
python-dotenv==1.0.1
python-Levenshtein==0.26.0
PyYAML==6.0.2
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
shellingham==1.5.4
smart-open==7.0.5
spacy==3.8.2
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.5
srsly==2.4.8
sympy==1.13.3
thinc==8.3.2
tokenizers==0.15.2
torch==2.4.1
tqdm==4.66.5
transformers==4.36.2
triton==3.0.0
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.4
wrapt==1.16.0